requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0
coverage==4.5.4           # via pytest-cov
importlib-metadata==0.23  # via pluggy, pytest
more-itertools==7.2.0     # via pytest, zipp
packaging==19.2           # via pytest
pluggy==0.13.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.4          # via packaging
pytest-cov==2.8.1
pytest==5.2.2
six==1.13.0               # via packaging
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
